asn1: Avoid undefined shifts by adding explicit cast
The error was:
asn1.c:681:23: runtime error: left shift of 255 by 24 places cannot be represented in type 'int'
	    #0 0x5e9b11 in decode_bit_field opensc/src/libopensc/asn1.c:681:23

https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=18487

+++ src/libopensc/asn1.c
@@ -678,7 +678,7 @@ static int decode_bit_field(const u8 * inbuf, size_t inlen, void *outbuf, size_t
 		return n;
 
 	for (i = 0; i < n; i += 8) {
-		field |= (data[i/8] << i);
+		field |= ((unsigned int) data[i/8] << i);
 	}
 	memcpy(outbuf, &field, outlen);
 	return 0;
